+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+BREAKING CHANGES:
+
+- user data are not gzipped anymore. If you want to gzip user data, you may want to use the cloudinit_config datasource.
+
 BUG FIX:
 
 - datasource `exoscale_instance_pool_list`: fix panic when instance pool with labels is found